+++ kymatio/scattering1d/backend/numpy_backend.py
@@ -1,30 +1,12 @@
 # Authors: Edouard Oyallon, Joakim Anden, Mathieu Andreux
 from collections import namedtuple
-from scipy.fftpack import fft, ifft
+import scipy
 import numpy as np
 
 BACKEND_NAME = 'numpy'
 
-
-def modulus_complex(x):
-    """Compute the complex modulus
-    Computes the modulus of x and stores the result in a complex tensor of the
-    same size, with the real part equal to the modulus and the imaginary part
-    equal to zero.
-    Parameters
-    ----------
-    x : tensor
-        A complex tensor (that is, whose last dimension is equal to 2).
-    Returns
-    -------
-    norm : tensor
-        A tensor with the same dimensions as x, such that norm[..., 0] contains
-        the complex modulus of x, while norm[..., 1] = 0.
-    """
-
-    norm = np.abs(x)
-
-    return norm
+from ...backend.numpy_backend import modulus, cdgmm, real
+from ...backend.base_backend import FFT
 
 
 def subsample_fourier(x, k):
@@ -129,53 +111,6 @@ def unpad(x, i0, i1):
     return x[..., i0:i1]
 
 
-def real(x):
-    """Real part of complex tensor
-    Takes the real part of a complex tensor, where the last axis corresponds
-    to the real and imaginary parts.
-    Parameters
-    ----------
-    x : tensor
-        A complex tensor (that is, whose last dimension is equal to 2).
-    Returns
-    -------
-    x_real : tensor
-        The tensor x[..., 0] which is interpreted as the real part of x.
-    """
-    return np.real(x)
-
-
-def fft1d_c2c(x, axis=-1):
-    """Compute the 1D FFT of a complex signal
-    Input
-    -----
-    x : tensor
-        A tensor of size (..., T, 2), where x[..., 0] is the real part and
-        x[..., 1] is the imaginary part.
-    Returns
-    -------
-    x_f : tensor
-        A tensor of the same size as x containing its Fourier transform in the
-        standard PyTorch FFT ordering.
-    """
-    return fft(x, axis=axis)
-
-
-def ifft1d_c2c(x, axis=-1):
-    """Compute the normalized 1D inverse FFT of a complex signal
-    Input
-    -----
-    x_f : tensor
-        A tensor of size (..., T, 2), where x_f[..., 0] is the real part and
-        x[..., 1] is the imaginary part. The frequencies are assumed to be in
-        the standard PyTorch FFT ordering.
-    Returns
-    -------
-    x : tensor
-        A tensor of the same size of x_f containing the normalized inverse
-        Fourier transform of x_f.
-    """
-    return ifft(x, axis=axis)
 
 
 def concatenate(arrays):
@@ -184,14 +119,17 @@ def concatenate(arrays):
 
 backend = namedtuple('backend',
                      ['name', 'modulus_complex', 'subsample_fourier', 'real',
-                      'unpad', 'fft1d_c2c', 'ifft1d_c2c', 'concatenate'])
+                      'unpad', 'fft', 'concatenate', 'cdgmm'])
 backend.name = 'numpy'
-backend.modulus_complex = modulus_complex
+backend.modulus_complex = modulus
 backend.subsample_fourier = subsample_fourier
 backend.real = real
 backend.unpad = unpad
 backend.pad = pad
 backend.pad_1d = pad_1d
-backend.fft1d_c2c = fft1d_c2c
-backend.ifft1d_c2c = ifft1d_c2c
+backend.cdgmm = cdgmm
+backend.fft = FFT(lambda x:scipy.fftpack.fft(x),
+                  lambda x:scipy.fftpack.ifft(x),
+                  lambda x:np.real(scipy.fftpack.ifft(x)),
+                  lambda x:None)
 backend.concatenate = concatenate
