+++ kymatio/scattering1d/backend/torch_backend.py
@@ -2,135 +2,13 @@
 
 import torch
 import torch.nn.functional as F
-from torch.autograd import Function
 
 from collections import namedtuple
 
 BACKEND_NAME = 'torch'
 
-def _is_complex(input):
-    return input.shape[-1] == 2
-
-class ModulusStable(Function):
-    """Stable complex modulus
-
-    This class implements a modulus transform for complex numbers which is
-    stable with respect to very small inputs (z close to 0), avoiding
-    returning nans in all cases.
-
-    Usage
-    -----
-    modulus = ModulusStable.apply  # apply inherited from Function
-    x_mod = modulus(x)
-
-    Parameters
-    ---------
-    x : tensor
-        The complex tensor (i.e., whose last dimension is two) whose modulus
-        we want to compute.
-
-    Returns
-    -------
-    output : tensor
-        A tensor of same size as the input tensor, except for the last
-        dimension, which is removed. This tensor is differentiable with respect
-        to the input in a stable fashion (so gradent of the modulus at zero is
-        zero).
-    """
-
-    @staticmethod
-    def forward(ctx, x):
-        """Forward pass of the modulus.
-
-        This is a static method which does not require an instantiation of the
-        class.
-
-        Arguments
-        ---------
-        ctx : context object
-            Collected during the forward pass. These are automatically added
-            by PyTorch and should not be touched. They are then used for the
-            backward pass.
-        x : tensor
-            The complex tensor whose modulus is to be computed.
-
-        Returns
-        -------
-        output : tensor
-            This contains the modulus computed along the last axis, with that
-            axis removed.
-        """
-        ctx.p = 2
-        ctx.dim = -1
-        ctx.keepdim = False
-
-        output = (x[...,0]*x[...,0] + x[...,1]*x[...,1]).sqrt()
-
-        ctx.save_for_backward(x, output)
-        return output
-
-    @staticmethod
-    def backward(ctx, grad_output):
-        """Backward pass of the modulus
-
-        This is a static method which does not require an instantiation of the
-        class.
-
-        Arguments
-        ---------
-        ctx : context object
-            Collected during the forward pass. These are automatically added
-            by PyTorch and should not be touched. They are then used for the
-            backward pass.
-        grad_output : tensor
-            The gradient with respect to the output tensor computed at the
-            forward pass.
-
-        Returns
-        -------
-        grad_input : tensor
-            The gradient with respect to the input.
-        """
-        x, output = ctx.saved_tensors
-        if ctx.dim is not None and ctx.keepdim is False and x.dim() != 1:
-            grad_output = grad_output.unsqueeze(ctx.dim)
-            output = output.unsqueeze(ctx.dim)
-
-        grad_input = x.mul(grad_output).div(output)
-
-        # Special case at 0 where we return a subgradient containing 0
-        grad_input.masked_fill_(output == 0, 0)
-
-        return grad_input
-
-# shortcut for ModulusStable.apply
-modulus = ModulusStable.apply
-
-def modulus_complex(x):
-    """Compute the complex modulus
-
-    Computes the modulus of x and stores the result in a complex tensor of the
-    same size, with the real part equal to the modulus and the imaginary part
-    equal to zero.
-
-    Parameters
-    ----------
-    x : tensor
-        A complex tensor (that is, whose last dimension is equal to 2).
-
-    Returns
-    -------
-    norm : tensor
-        A tensor with the same dimensions as x, such that norm[..., 0] contains
-        the complex modulus of x, while norm[..., 1] = 0.
-    """
-    if not _is_complex(x):
-        raise TypeError('The input should be complex.')
-
-    norm = torch.zeros_like(x)
-    norm[...,0] = modulus(x)
-
-    return norm
+from ...backend.torch_backend import _is_complex, Modulus, concatenate, type_checks, cdgmm, real
+from ...backend.base_backend import FFT
 
 def subsample_fourier(x, k):
     """Subsampling in the Fourier domain
@@ -252,72 +130,21 @@ def unpad(x, i0, i1):
     """
     return x[..., i0:i1]
 
-def real(x):
-    """Real part of complex tensor
 
-    Takes the real part of a complex tensor, where the last axis corresponds
-    to the real and imaginary parts.
+fft = FFT(lambda x: torch.fft(x, 1, normalized=False),
+    lambda x: torch.ifft(x, 1, normalized=False),
+    lambda x: torch.irfft(x, 1, normalized=False, onesided=False),
+    type_checks)
 
-    Parameters
-    ----------
-    x : tensor
-        A complex tensor (that is, whose last dimension is equal to 2).
 
-    Returns
-    -------
-    x_real : tensor
-        The tensor x[..., 0] which is interpreted as the real part of x.
-    """
-    return x[..., 0]
-
-def fft1d_c2c(x):
-    """Compute the 1D FFT of a complex signal
-
-    Input
-    -----
-    x : tensor
-        A tensor of size (..., T, 2), where x[..., 0] is the real part and
-        x[..., 1] is the imaginary part.
-
-    Returns
-    -------
-    x_f : tensor
-        A tensor of the same size as x containing its Fourier transform in the
-        standard PyTorch FFT ordering.
-    """
-    return torch.fft(x, signal_ndim=1)
-
-def ifft1d_c2c(x):
-    """Compute the normalized 1D inverse FFT of a complex signal
-
-    Input
-    -----
-    x_f : tensor
-        A tensor of size (..., T, 2), where x_f[..., 0] is the real part and
-        x[..., 1] is the imaginary part. The frequencies are assumed to be in
-        the standard PyTorch FFT ordering.
-
-    Returns
-    -------
-    x : tensor
-        A tensor of the same size of x_f containing the normalized inverse
-        Fourier transform of x_f.
-    """
-    return torch.ifft(x, signal_ndim=1)
-
-def concatenate(arrays):
-    return torch.stack(arrays, dim=-2)
-
-backend = namedtuple('backend', ['name', 'modulus_complex', 'subsample_fourier', 'real', 'unpad', 'fft1d_c2c',\
-                                 'ifft1d_c2c', 'concatenate'])
+backend = namedtuple('backend', ['name', 'modulus_complex', 'subsample_fourier', 'real', 'unpad', 'fft', 'concatenate'])
 backend.name = 'torch'
-backend.modulus_complex = modulus_complex
-backend.ModulusStable = ModulusStable
+backend.modulus_complex = Modulus()
 backend.subsample_fourier = subsample_fourier
 backend.real = real
 backend.unpad = unpad
+backend.cdgmm = cdgmm
 backend.pad = pad
 backend.pad_1d = pad_1d
-backend.fft1d_c2c = fft1d_c2c
-backend.ifft1d_c2c = ifft1d_c2c
-backend.concatenate = concatenate
+backend.fft = fft
+backend.concatenate = lambda x: concatenate(x, -2)
