+++ kymatio/scattering1d/backend/torch_skcuda_backend.py
@@ -219,19 +219,17 @@ def subsample_fourier(x, k):
     return subsamplefourier(x,k)
 
 
-from .torch_backend import ifft1d_c2c, fft1d_c2c, real, unpad, pad, pad_1d, concatenate, ModulusStable
+from .torch_backend import real, cdgmm, unpad, pad, pad_1d, concatenate, fft
 
-backend = namedtuple('backend', ['name', 'modulus_complex', 'subsample_fourier', 'real', 'unpad', 'fft1d_c2c',
-                                 'ifft1d_c2c', 'concatenate'])
+backend = namedtuple('backend', ['name', 'modulus_complex', 'subsample_fourier', 'real', 'unpad', 'fft', 'concatenate'])
 
 backend.name = 'torch_skcuda'
 backend.modulus_complex = modulus_complex
-backend.ModulusStable = ModulusStable
 backend.subsample_fourier = subsample_fourier
 backend.real = real
+backend.cdgmm = cdgmm
 backend.unpad = unpad
 backend.pad = pad
 backend.pad_1d = pad_1d
-backend.fft1d_c2c = fft1d_c2c
-backend.ifft1d_c2c = ifft1d_c2c
-backend.concatenate = concatenate
+backend.fft = fft
+backend.concatenate = lambda x: concatenate(x, -2)
