+++ kymatio/scattering1d/core/scattering1d.py
@@ -58,9 +58,9 @@ def scattering1d(x, pad, unpad, backend, J, psi1, psi2, phi, pad_left=0, pad_rig
     """
     subsample_fourier = backend.subsample_fourier
     modulus_complex = backend.modulus_complex
-    fft1d_c2c = backend.fft1d_c2c
-    ifft1d_c2c = backend.ifft1d_c2c
     real = backend.real
+    fft = backend.fft
+    cdgmm = backend.cdgmm
     concatenate = backend.concatenate
 
 
@@ -77,16 +77,15 @@ def scattering1d(x, pad, unpad, backend, J, psi1, psi2, phi, pad_left=0, pad_rig
     U_0 = pad(x, pad_left=pad_left, pad_right=pad_right)
 
     # compute the Fourier transform
-    U_0_hat = fft1d_c2c(U_0)
+    U_0_hat = fft(U_0, 'C2C')
 
     # Get S0
     k0 = max(J - oversampling, 0)
 
     if average:
-        S_0_c = U_0_hat * phi[0]
+        S_0_c = cdgmm(U_0_hat, phi[0])
         S_0_hat = subsample_fourier(S_0_c, 2**k0)
-        S_0_c = ifft1d_c2c(S_0_hat)
-        S_0_r = real(S_0_c)
+        S_0_r = fft(S_0_hat, 'C2R', inverse=True)
 
         S_0 = unpad(S_0_r, ind_start[k0], ind_end[k0])
     else:
@@ -106,24 +105,23 @@ def scattering1d(x, pad, unpad, backend, J, psi1, psi2, phi, pad_left=0, pad_rig
 
         assert psi1[n1]['xi'] < 0.5 / (2**k1)
 
-        U_1_c = U_0_hat * psi1[n1][0]
+        U_1_c = cdgmm(U_0_hat, psi1[n1][0])
         U_1_hat = subsample_fourier(U_1_c, 2**k1)
-        U_1_c = ifft1d_c2c(U_1_hat)
+        U_1_c = fft(U_1_hat, 'C2C', inverse=True)
 
         # Take the modulus
         U_1_m = modulus_complex(U_1_c)
 
         if average or max_order > 1:
-            U_1_hat = fft1d_c2c(U_1_m)
+            U_1_hat = fft(U_1_m, 'C2C')
 
         if average:
             # Convolve with phi_J
             k1_J = max(J - k1 - oversampling, 0)
 
-            S_1_c = U_1_hat * phi[k1]
+            S_1_c = cdgmm(U_1_hat, phi[k1])
             S_1_hat = subsample_fourier(S_1_c, 2**k1_J)
-            S_1_c = ifft1d_c2c(S_1_hat)
-            S_1_r = real(S_1_c)
+            S_1_r = fft(S_1_hat, 'C2R', inverse=True)
 
             S_1 = unpad(S_1_r, ind_start[k1_J + k1], ind_end[k1_J + k1])
         else:
@@ -148,23 +146,22 @@ def scattering1d(x, pad, unpad, backend, J, psi1, psi2, phi, pad_left=0, pad_rig
                     # convolution + downsampling
                     k2 = max(j2 - k1 - oversampling, 0)
 
-                    U_2_c = U_1_hat * psi2[n2][k1]
+                    U_2_c = cdgmm(U_1_hat, psi2[n2][k1])
                     U_2_hat = subsample_fourier(U_2_c, 2**k2)
                     # take the modulus
-                    U_2_c = ifft1d_c2c(U_2_hat)
+                    U_2_c = fft(U_2_hat, 'C2C', inverse=True)
 
                     U_2_m = modulus_complex(U_2_c)
 
                     if average:
-                        U_2_hat = fft1d_c2c(U_2_m)
+                        U_2_hat = fft(U_2_m, 'C2C')
 
                         # Convolve with phi_J
                         k2_J = max(J - k2 - k1 - oversampling, 0)
 
-                        S_2_c = U_2_hat * phi[k1 + k2]
+                        S_2_c = cdgmm(U_2_hat, phi[k1 + k2])
                         S_2_hat = subsample_fourier(S_2_c, 2**k2_J)
-                        S_2_c = ifft1d_c2c(S_2_hat)
-                        S_2_r = real(S_2_c)
+                        S_2_r = fft(S_2_hat, 'C2R', inverse=True)
 
                         S_2 = unpad(S_2_r, ind_start[k1 + k2 + k2_J], ind_end[k1 + k2 + k2_J])
                     else:
